kernel-3.0.x: update hw_nat to 2.50.x, fix HW_VLAN_TX, fix IPv6 offload for 1xRGMII, forward all mcast and bcast traffic to CPU, code refactoring

+++ trunk/linux-3.0.x/net/nat/hw_nat/ac_ioctl.h
@@ -55,7 +55,7 @@ struct ac_args {
 	unsigned long ip_s;	/* Start Ip Address */
 	unsigned long ip_e;	/* End Ip Address */
 	unsigned int ag_idx;	/* account group index */
-	unsigned int cnt;	/* pkt_cnt or byt_cnt */
+	long long cnt;		/* pkt_cnt or byt_cnt */
 	enum AcResult result;
 };
